+++ ReAwakn_App/next_js_app/src/app/meetings/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React, { useState, useEffect } from "react";
-import { useRouter } from "next/navigation";
+import { useRouter, useSearchParams } from "next/navigation";
 import { useUserTimezone } from "@/hooks/useUserTimezone";
 import { Calendar, momentLocalizer } from "react-big-calendar";
 import moment from "moment-timezone";
@@ -8,20 +8,67 @@ import "react-big-calendar/lib/css/react-big-calendar.css";
 import { supabase } from "@/app/utils/supabase/client";
 import { getAuthUser } from "@/utility_methods/userUtils";
 import MeetingDetailsModal from "./MeetingDetailsModal";
-import { Meeting, CalendarMeeting } from "@/types/types";
+import {
+  Meeting,
+  CalendarMeeting,
+  CalendarUser,
+  RankedCalendarEvent,
+} from "@/types/types";
+import MeetingConfirmationModal from "../schedule_meeting/MeetingConfirmationModal";
+import {
+  findOverlappingTimeSlots,
+  convertToCalendarEvents,
+} from "@/utility_methods/schedulingUtils";
+import {
+  filterAvailableEvents,
+  prepareUserForRanking,
+  prepareSlotsForRanking,
+  markTopMeetingSlots,
+} from "@/utility_methods/meetingUtils";
+import {
+  calculateUserSimilarityScores,
+  areFriends,
+} from "@/utility_methods/memberCardUtils";
 
 const localizer = momentLocalizer(moment);
 
 export default function MeetingsPage() {
   const router = useRouter();
+  const searchParams = useSearchParams();
+  const targetUserId = searchParams.get("userId");
+
   const [meetings, setMeetings] = useState<CalendarMeeting[]>([]);
   const [selectedMeeting, setSelectedMeeting] =
     useState<CalendarMeeting | null>(null);
-  const [isModalOpen, setIsModalOpen] = useState(false);
+  const [isDetailsModalOpen, setIsDetailsModalOpen] = useState(false);
+
+  const [loggedInUser, setLoggedInUser] = useState<CalendarUser | null>(null);
+  const [targetUser, setTargetUser] = useState<CalendarUser | null>(null);
+  const [availableSlots, setAvailableSlots] = useState<
+    RankedCalendarEvent[] | null
+  >([]);
+  const [selectedSlot, setSelectedSlot] = useState<RankedCalendarEvent | null>(
+    null
+  );
+  const [isScheduleModalOpen, setIsScheduleModalOpen] = useState(false);
+  const [teachingHours, setTeachingHours] = useState<number>(0);
+  const [matchingSkill, setMatchingSkill] = useState<string>("");
+  const [usersAreFriends, setUsersAreFriends] = useState<boolean>(false);
+  const [checkingFriendship, setCheckingFriendship] = useState<boolean>(false);
+
   const [isLoading, setIsLoading] = useState(true);
-  const [loggedInUserData, setLoggedInUserData] = useState(null);
   const [currentDate, setCurrentDate] = useState(new Date());
-  const userTimezone = useUserTimezone(loggedInUserData);
+  const [isSchedulingMode, setIsSchedulingMode] = useState(false);
+  const userTimezone = useUserTimezone(loggedInUser || null);
+  useEffect(() => {
+    if (targetUserId) {
+      setIsSchedulingMode(true);
+      setCheckingFriendship(true);
+    } else {
+      setIsSchedulingMode(false);
+    }
+  }, [targetUserId]);
+
   const fetchMeetings = async () => {
     setIsLoading(true);
     try {
@@ -88,7 +135,43 @@ export default function MeetingsPage() {
             .eq("id", user.id)
             .single();
 
-          setLoggedInUserData(data);
+          setLoggedInUser(data);
+
+          if (targetUserId && isSchedulingMode) {
+            const { data: targetUserData } = await supabase
+              .from("users")
+              .select("*")
+              .eq("id", targetUserId)
+              .single();
+
+            setTargetUser(targetUserData);
+
+            const friendshipStatus = await areFriends(user.id, targetUserId);
+            setUsersAreFriends(friendshipStatus);
+            setCheckingFriendship(false);
+
+            if (friendshipStatus && data && targetUserData) {
+              try {
+                const result = await calculateUserSimilarityScores(
+                  data.id,
+                  targetUserData.id
+                );
+
+                const { teaching_hours, matching_skill } = result;
+
+                if (
+                  teaching_hours !== undefined &&
+                  teaching_hours > 0 &&
+                  matching_skill
+                ) {
+                  setTeachingHours(teaching_hours);
+                  setMatchingSkill(matching_skill);
+                }
+              } catch (error) {
+                console.error("Error getting teaching hours", error);
+              }
+            }
+          }
         }
       } catch (error) {
         alert(error);
@@ -97,11 +180,158 @@ export default function MeetingsPage() {
 
     fetchUserData();
     fetchMeetings();
-  }, []);
+  }, [targetUserId, isSchedulingMode]);
 
-  const selectMeeting = (event: CalendarMeeting) => {
-    setSelectedMeeting(event);
-    setIsModalOpen(true);
+  useEffect(() => {
+    const findAvailableSlots = async () => {
+      if (
+        !isSchedulingMode ||
+        !loggedInUser ||
+        !targetUser ||
+        !userTimezone ||
+        !usersAreFriends
+      ) {
+        return;
+      }
+
+      const slots = await findAvailableMeetingSlots(
+        loggedInUser.id,
+        targetUser.id,
+        teachingHours
+      );
+      setAvailableSlots(slots);
+    };
+
+    findAvailableSlots();
+  }, [
+    loggedInUser,
+    targetUser,
+    userTimezone,
+    teachingHours,
+    isSchedulingMode,
+    usersAreFriends,
+  ]);
+
+  const findAvailableMeetingSlots = async (
+    hostId: string,
+    targetId: string,
+    hours: number = teachingHours
+  ) => {
+    const { data: hostData } = await supabase
+      .from("users")
+      .select("*")
+      .eq("id", hostId)
+      .single();
+
+    const { data: targetData } = await supabase
+      .from("users")
+      .select("*")
+      .eq("id", targetId)
+      .single();
+
+    if (!hostData?.availability || !targetData?.availability) {
+      return [];
+    }
+
+    try {
+      const hostAvailabilityArray = JSON.parse(hostData.availability);
+      const targetAvailabilityArray = JSON.parse(targetData.availability);
+
+      const overlappingSlots = findOverlappingTimeSlots(
+        hostAvailabilityArray,
+        targetAvailabilityArray
+      );
+
+      const calendarEvents = convertToCalendarEvents(
+        overlappingSlots,
+        userTimezone || "UTC"
+      );
+
+      const startDate = moment().startOf("day").toISOString();
+      const endDate = moment().add(2, "months").endOf("month").toISOString();
+
+      const { data: existingMeetings, error: meetingsError } = await supabase
+        .from("meetings")
+        .select("*")
+        .or(
+          `host_id.eq.${hostId},guest_id.eq.${hostId},host_id.eq.${targetId},guest_id.eq.${targetId}`
+        )
+        .gte("start_time", startDate)
+        .lte("start_time", endDate);
+
+      if (meetingsError) {
+        console.error("Error fetching existing meetings", meetingsError);
+      }
+
+      const availableEvents = filterAvailableEvents(
+        calendarEvents,
+        existingMeetings || []
+      );
+
+      if (availableEvents.length === 0) {
+        return [];
+      }
+
+      const host = prepareUserForRanking(hostId, hostData);
+      const target = prepareUserForRanking(targetId, targetData);
+
+      const slotsForRanking = prepareSlotsForRanking(availableEvents);
+
+      try {
+        const response = await fetch("/meeting_ranker_api", {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            user1: host,
+            user2: target,
+            slots: slotsForRanking,
+            existingMeetings: existingMeetings || [],
+          }),
+        });
+
+        if (!response.ok) {
+          return availableEvents;
+        }
+
+        const { top_slots } = await response.json();
+
+        if (top_slots && top_slots.length > 0) {
+          const numSlotsToShow = hours > 0 ? hours : 5;
+          return markTopMeetingSlots(
+            availableEvents,
+            top_slots,
+            numSlotsToShow
+          ) as RankedCalendarEvent[];
+        } else {
+          return availableEvents as RankedCalendarEvent[];
+        }
+      } catch (error) {
+        return availableEvents;
+      }
+    } catch (error) {
+      console.error("Error finding available meeting slots", error);
+      return [];
+    }
+  };
+
+  const selectEvent = (event: any) => {
+    if (event.id) {
+      setSelectedMeeting(event);
+      setIsDetailsModalOpen(true);
+    } else {
+      const oneHourSlot = {
+        start: event.start,
+        end: event.end,
+        startUTC: event.startUTC || "",
+        endUTC: event.endUTC || "",
+        action: "click",
+      };
+
+      setSelectedSlot(oneHourSlot);
+      setIsScheduleModalOpen(true);
+    }
   };
 
   const acceptMeeting = async (meetingId: string) => {
@@ -127,7 +357,7 @@ export default function MeetingsPage() {
       await supabase.from("meetings").delete().eq("id", meetingId);
 
       if (selectedMeeting && selectedMeeting.id === meetingId) {
-        setIsModalOpen(false);
+        setIsDetailsModalOpen(false);
         setSelectedMeeting(null);
       }
 
@@ -139,22 +369,97 @@ export default function MeetingsPage() {
     }
   };
 
-  const calendarCSS = (event: CalendarMeeting) => {
-    let backgroundColor = event.is_confirmed ? "#4CAF50" : "#FFC107";
+  const bookMeeting = async ({ title }: { title: string }) => {
+    try {
+      if (!loggedInUser || !targetUser || !selectedSlot) {
+        throw new Error("Missing required information to book meeting");
+      }
+
+      const startTimeISO = selectedSlot.startUTC
+        ? moment.utc(selectedSlot.startUTC).toISOString()
+        : moment.utc(selectedSlot.start).toISOString();
+
+      const { data: existingMeetings, error: conflictError } = await supabase
+        .from("meetings")
+        .select("*")
+        .or(
+          `host_id.eq.${loggedInUser.id},guest_id.eq.${loggedInUser.id},host_id.eq.${targetUser.id},guest_id.eq.${targetUser.id}`
+        )
+        .eq("start_time", startTimeISO);
+
+      if (conflictError) {
+        alert("Error checking for meeting conflicts");
+        throw new Error("Error checking for meeting conflicts");
+      }
+
+      if (existingMeetings && existingMeetings.length > 0) {
+        alert(
+          "This time slot is no longer available. Please select another time."
+        );
+        return;
+      }
+
+      const { error: insertError } = await supabase.from("meetings").insert({
+        host_id: loggedInUser.id,
+        guest_id: targetUser.id,
+        start_time: startTimeISO,
+        title: title,
+        is_confirmed: false,
+      });
+
+      if (insertError) {
+        alert("Failed to schedule meeting");
+        throw new Error("Failed to schedule meeting");
+      }
+
+      alert("Meeting scheduled successfully!");
+      setIsScheduleModalOpen(false);
+      fetchMeetings();
+    } catch (error) {
+      alert("Failed to schedule meeting. Please try again.");
+    }
+  };
+
+  const calendarCSS = (event: any) => {
+    if (event.id) {
+      let backgroundColor = event.is_confirmed ? "#4CAF50" : "#FFC107";
+
+      return {
+        style: {
+          backgroundColor,
+          borderRadius: "5px",
+          opacity: 0.8,
+          color: "white",
+          border: "0px",
+          display: "block",
+        },
+      };
+    }
+
+    if (event.resource?.isBestSlot) {
+      return {
+        style: {
+          backgroundColor: "#FF5722",
+          cursor: "pointer",
+          border: "2px solid #D84315",
+          boxShadow: "0 0 5px rgba(255, 87, 34, 0.5)",
+          borderRadius: "5px",
+          color: "white",
+        },
+      };
+    }
 
     return {
       style: {
-        backgroundColor,
+        backgroundColor: "#78909C",
+        cursor: "pointer",
         borderRadius: "5px",
-        opacity: 0.8,
         color: "white",
-        border: "0px",
-        display: "block",
       },
     };
   };
 
-  if (isLoading) {
+  if (isLoading || (isSchedulingMode && checkingFriendship)) {
     return (
       <div className="flex justify-center items-center min-h-screen">
         <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-500"></div>
@@ -162,9 +467,56 @@ export default function MeetingsPage() {
     );
   }
 
+  if (isSchedulingMode && !usersAreFriends && targetUser) {
+    return (
+      <div className="container mx-auto px-4 py-8">
+        <div className="bg-white rounded-lg shadow-md p-8 text-center">
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            className="h-16 w-16 text-red-500 mx-auto mb-4"
+            fill="none"
+            viewBox="0 0 24 24"
+            stroke="currentColor"
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth={2}
+              d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z"
+            />
+          </svg>
+          <h2 className="text-2xl font-bold mb-4">Connection Required</h2>
+          <p className="text-gray-600 mb-6">
+            You need to be connected with{" "}
+            {targetUser?.name || targetUser?.display_name} before you can
+            schedule a meeting.
+          </p>
+          <div className="flex justify-center">
+            <button
+              onClick={() => router.push("/")}
+              className="px-6 py-2 bg-blue-500 hover:bg-blue-600 text-white font-medium rounded-md transition-colors duration-200"
+            >
+              Back to Home
+            </button>
+          </div>
+        </div>
+      </div>
+    );
+  }
+
+  const allEvents = isSchedulingMode
+    ? [...meetings, ...(availableSlots || [])]
+    : meetings;
+
   return (
     <div className="container mx-auto px-4 py-8 text-black">
-      <h1 className="text-2xl font-bold mb-4 text-black">My Meetings</h1>
+      <h1 className="text-2xl font-bold mb-4 text-black">
+        {isSchedulingMode
+          ? `Schedule a Meeting with ${
+              targetUser?.name || targetUser?.display_name
+            }`
+          : "My Meetings"}
+      </h1>
 
       <div className="bg-blue-50 border border-blue-200 rounded-md p-3 mb-4 flex items-center">
         <svg
@@ -186,10 +538,50 @@ export default function MeetingsPage() {
         </div>
       </div>
 
+      {isSchedulingMode && teachingHours > 0 && matchingSkill && (
+        <div className="bg-green-50 border border-green-200 rounded-md p-3 mb-4">
+          <div className="flex items-center">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              className="h-5 w-5 text-green-500 mr-2"
+              viewBox="0 0 20 20"
+              fill="currentColor"
+            >
+              <path
+                fillRule="evenodd"
+                d="M10 18a8 8 0 100-16 8 8 0 000 16zm1-12a1 1 0 10-2 0v4a1 1 0 00.293.707l2.828 2.829a1 1 0 101.415-1.415L11 9.586V6z"
+                clipRule="evenodd"
+              />
+            </svg>
+            <p className="text-sm text-green-700 font-medium">
+              Teaching Plan Information
+            </p>
+          </div>
+          <div className="mt-2 text-sm text-green-700">
+            <p>
+              <span className="font-medium">Matching Skill:</span>{" "}
+              {matchingSkill}
+            </p>
+            <p>
+              <span className="font-medium">Teaching Hours Required:</span>{" "}
+              {teachingHours} hours
+            </p>
+          </div>
+        </div>
+      )}
+
       <div className="bg-white rounded-lg shadow-md p-6 mb-6">
         <div className="flex justify-between items-center mb-4">
           <h2 className="text-lg font-semibold text-black">Meeting Calendar</h2>
           <div className="flex space-x-2">
+            {isSchedulingMode && (
+              <button
+                onClick={() => router.push("/meetings")}
+                className="px-4 py-2 bg-gray-100 hover:bg-gray-200 text-gray-800 text-sm font-medium rounded-md transition-colors duration-200 focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2"
+              >
+                View My Meetings
+              </button>
+            )}
             <button
               onClick={() => router.push("/")}
               className="px-4 py-2 bg-blue-500 hover:bg-blue-600 text-white text-sm font-medium rounded-md transition-colors duration-200 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2"
@@ -204,20 +596,32 @@ export default function MeetingsPage() {
             <div className="w-4 h-4 bg-[#4CAF50] rounded-full mr-2"></div>
             <span className="text-sm text-black">Confirmed</span>
           </div>
-          <div className="flex items-center">
+          <div className="flex items-center mr-4">
             <div className="w-4 h-4 bg-[#FFC107] rounded-full mr-2"></div>
             <span className="text-sm text-black">Pending</span>
           </div>
+          {isSchedulingMode && (
+            <>
+              <div className="flex items-center mr-4">
+                <div className="w-4 h-4 bg-[#FF5722] rounded mr-2"></div>
+                <span className="text-sm text-black">Best Meeting Time</span>
+              </div>
+              <div className="flex items-center">
+                <div className="w-4 h-4 bg-[#78909C] rounded mr-2"></div>
+                <span className="text-sm text-black">Available Time</span>
+              </div>
+            </>
+          )}
         </div>
 
         <div className="border border-gray-200 rounded-md">
           <div className="h-[700px] text-black">
             <Calendar
               localizer={localizer}
-              events={meetings}
+              events={allEvents}
               startAccessor="start"
               endAccessor="end"
-              onSelectEvent={selectMeeting}
+              onSelectEvent={selectEvent}
               eventPropGetter={calendarCSS}
               defaultView="week"
               views={["month", "week", "day"]}
@@ -227,12 +631,14 @@ export default function MeetingsPage() {
               max={new Date(0, 0, 0, 23, 0)}
               formats={{
                 timeGutterFormat: (date: Date) =>
-                  moment(date).tz(userTimezone).format("h A"),
+                  moment(date)
+                    .tz(userTimezone || "UTC")
+                    .format("h A"),
                 eventTimeRangeFormat: (range) => {
                   return `${moment(range.start)
-                    .tz(userTimezone)
+                    .tz(userTimezone || "UTC")
                     .format("h:mm A")} - ${moment(range.end)
-                    .tz(userTimezone)
+                    .tz(userTimezone || "UTC")
                     .format("h:mm A")}`;
                 },
               }}
@@ -243,11 +649,23 @@ export default function MeetingsPage() {
 
       {selectedMeeting && (
         <MeetingDetailsModal
-          isOpen={isModalOpen}
-          onClose={() => setIsModalOpen(false)}
+          isOpen={isDetailsModalOpen}
+          onClose={() => setIsDetailsModalOpen(false)}
           meeting={selectedMeeting}
           onConfirm={acceptMeeting}
           onCancel={cancelMeeting}
+        />
+      )}
+
+      {isSchedulingMode && selectedSlot && targetUser && (
+        <MeetingConfirmationModal
+          isOpen={isScheduleModalOpen}
+          onClose={() => setIsScheduleModalOpen(false)}
+          selectedSlot={selectedSlot}
+          targetUser={targetUser}
+          onConfirm={bookMeeting}
+          teachingHours={teachingHours}
+          matchingSkill={matchingSkill}
         />
       )}
     </div>
