+++ ReAwakn_App/next_js_app/src/app/search_page/MemberCard.tsx
@@ -1,4 +1,4 @@
-import { MemberCardProps, MemberWithSimilarity } from "@/types/types";
+import { MemberCardProps } from "@/types/types";
 import {
   getSimilarityColor,
   getSimilarityLabel,
